app/requirements-dev.txt
# PyOxynet Development Dependencies
# Scientific API modernization - Phase 1: Foundation & Security
# IMPORTANT: Install with --extra-index-url https://google-coral.github.io/py-repo/

--extra-index-url https://google-coral.github.io/py-repo/

# Production dependencies (include all from requirements.txt)
-r requirements.txt

# Development and Testing Framework
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-flask>=1.3.0
pytest-mock>=3.12.0

# Code Quality
black>=23.0.0
flake8>=6.0.0
mypy>=1.8.0

# API Documentation and Validation
marshmallow>=3.20.0
apispec>=6.4.0

# Additional Scientific Libraries for Development
scipy>=1.11.0
pandas>=2.1.0

# Security and Privacy (Development)
bcrypt>=4.1.0

# Development Server and Tools
python-dotenv

# Async Processing (Future Phase 2)
celery>=5.3.0
redis>=5.0.0